wolfram@the-dreams.de:
When checking the commandline-help for generating the manpage from it,
I wondered if CONFIG_NONE is really needed. Removing it, makes IMHO
this upcoming task easier and also simplifies the code, as it has often
to be treated as an exception. I think the information presented there
can also be presented otherwise (like preceeding 'conf-var:').