Antonio Borneo <borneo.antonio@gmail.com>
there is a variable not initialized in vpnc-nortel branch.
Moving the code from setup_link() to Nortel specific do_phase2(),
s->ipsec.life.start initialization got lost.
This cause printing of wrong elapsed time information when flag
"--no-detach" is used together with debug level 2 or more.

The attached patch fix it in the code, in a reasonable place.