Freek Dijkstra <freek@macfreek.nl>
tested r457 + patch-20110815.diff on Mac OS 10.7 (x86_64).
That did not work. make failed to compile:

In file included from sysdep.c:71:
sysdep.h:240: error: conflicting types for ‘getline’
/usr/include/stdio.h:449: error: previous declaration of ‘getline’ was here

This is a known problem, described at
https://github.com/mxcl/homebrew/issues/6437
The issue is that vpnc does not check if getline is available, but
merely assumes it is not based on the presence of the __APPLE__ macro.
This works fine for 10.5 and 10.6, but fails for 10.7.
Rather then relying on the __APPLE__ macro (or more specific checks such
as __MAC_OS_X_VERSION_MIN_REQUIRED >= __MAC_10_7), would it be possible
to have pkg-config check for the getline() function in stdio?